+++ src/insiphy/__init__.py
@@ -1,3 +1,3 @@
 """INSIPHY: intragenic synteny phylogenetics."""
 
-__version__ = "0.4.0"
+__version__ = "0.6.0"